refactor: clean up deprecated ArticleSuggestions-Dialogflow snippets and remove unused AnswerRecordManagement class
