+++ dialogflow/snippets/src/main/java/com/example/dialogflow/ConversationProfileManagement.java
@@ -16,24 +16,12 @@
 
 package com.example.dialogflow;
 
-// [START dialogflow_create_conversation_profile_article_suggestion]
-
 import com.google.api.gax.rpc.ApiException;
 import com.google.cloud.dialogflow.v2.ConversationProfile;
 import com.google.cloud.dialogflow.v2.ConversationProfilesClient;
 import com.google.cloud.dialogflow.v2.CreateConversationProfileRequest;
-import com.google.cloud.dialogflow.v2.HumanAgentAssistantConfig;
-import com.google.cloud.dialogflow.v2.HumanAgentAssistantConfig.SuggestionConfig;
-import com.google.cloud.dialogflow.v2.HumanAgentAssistantConfig.SuggestionFeatureConfig;
-import com.google.cloud.dialogflow.v2.HumanAgentAssistantConfig.SuggestionQueryConfig;
-import com.google.cloud.dialogflow.v2.HumanAgentAssistantConfig.SuggestionQueryConfig.KnowledgeBaseQuerySource;
-import com.google.cloud.dialogflow.v2.HumanAgentAssistantConfig.SuggestionTriggerSettings;
-import com.google.cloud.dialogflow.v2.KnowledgeBaseName;
 import com.google.cloud.dialogflow.v2.LocationName;
-import com.google.cloud.dialogflow.v2.SuggestionFeature;
-import com.google.cloud.dialogflow.v2.SuggestionFeature.Type;
 import java.io.IOException;
-import java.util.Optional;
 
 public class ConversationProfileManagement {
 
@@ -45,79 +33,24 @@ public class ConversationProfileManagement {
     // Set display name of the new conversation profile
     String conversationProfileDisplayName = "my-conversation-profile-display-name";
 
-    // Set knowledge base id for Article Suggestion feature.
-    // See details about how to create a knowledge base here,
-    // https://cloud.google.com/agent-assist/docs/article-suggestion.
-    String articleSuggestionKnowledgeBaseId = "my-article-suggestion-knowledge-base-id";
-
     // Create a conversation profile
-    createConversationProfileArticleSuggestion(
-        projectId,
-        conversationProfileDisplayName,
-        location,
-        Optional.of(articleSuggestionKnowledgeBaseId));
+    createConversationProfile(projectId, conversationProfileDisplayName, location);
   }
 
-  // Set suggestion trigger with no_smalltalk and only_send_user both true, which means that
-  // the suggestion is not triggered if last utterance is small talk and is only triggered
-  // if participant role of last utterance is END_USER.
-  public static SuggestionTriggerSettings buildSuggestionTriggerSettings() {
-    return SuggestionTriggerSettings.newBuilder().setNoSmalltalk(true).setOnlyEndUser(true).build();
-  }
-
-  // Set the configuration for suggestion query, including the knowledge base query source
-  // and maximum number of results to return.
-  public static SuggestionQueryConfig buildSuggestionQueryConfig(
-      KnowledgeBaseName knowledgeBaseName) {
-    return SuggestionQueryConfig.newBuilder()
-        .setKnowledgeBaseQuerySource(
-            KnowledgeBaseQuerySource.newBuilder().addKnowledgeBases(knowledgeBaseName.toString()))
-        .setMaxResults(3)
-        .build();
-  }
-
-  // Create a conversation profile with given values about Article Suggestion.
-  public static void createConversationProfileArticleSuggestion(
+  // Create a basic conversation profile.
+  public static void createConversationProfile(
       String projectId,
       String displayName,
-      String location,
-      Optional<String> articleSuggestionKnowledgeBaseId)
-      throws ApiException, IOException {
+      String location) throws ApiException, IOException {
     try (ConversationProfilesClient conversationProfilesClient =
         ConversationProfilesClient.create()) {
-      // Create a builder for agent assistance configuration
-      SuggestionConfig.Builder suggestionConfigBuilder = SuggestionConfig.newBuilder();
-
-      // Add knowledge base for Article Suggestion feature
-      if (articleSuggestionKnowledgeBaseId.isPresent()) {
-        KnowledgeBaseName articleSuggestionKbName =
-            KnowledgeBaseName.of(projectId, articleSuggestionKnowledgeBaseId.get());
-
-        // Build configuration for Article Suggestion feature
-        SuggestionFeatureConfig articleSuggestionFeatureConfig =
-            SuggestionFeatureConfig.newBuilder()
-                .setSuggestionFeature(
-                    SuggestionFeature.newBuilder().setType(Type.ARTICLE_SUGGESTION).build())
-                .setSuggestionTriggerSettings(buildSuggestionTriggerSettings())
-                .setQueryConfig(buildSuggestionQueryConfig(articleSuggestionKbName))
-                .build();
-
-        // Add Article Suggestion feature to agent assistance configuration
-        suggestionConfigBuilder.addFeatureConfigs(articleSuggestionFeatureConfig);
-      }
-
       LocationName locationName = LocationName.of(projectId, location);
-      // Set a conversation profile with target configurations
       ConversationProfile targetConversationProfile =
           ConversationProfile.newBuilder()
               .setDisplayName(displayName)
               .setLanguageCode("en-US")
-              .setHumanAgentAssistantConfig(
-                  HumanAgentAssistantConfig.newBuilder()
-                      .setHumanAgentSuggestionConfig(suggestionConfigBuilder.build()))
               .build();
 
-      // Create a conversation profile
       ConversationProfile createdConversationProfile =
           conversationProfilesClient.createConversationProfile(
               CreateConversationProfileRequest.newBuilder()
@@ -131,4 +64,3 @@ public class ConversationProfileManagement {
     }
   }
 }
-// [END dialogflow_create_conversation_profile_article_suggestion]
