+++ dialogflow/snippets/src/test/java/com/example/dialogflow/CreateConversationTest.java
@@ -25,7 +25,6 @@ import com.google.cloud.dialogflow.v2.ConversationProfilesClient;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
-import java.util.Optional;
 import java.util.UUID;
 import org.junit.After;
 import org.junit.Before;
@@ -71,8 +70,8 @@ public class CreateConversationTest {
     System.setOut(newOutputStream);
 
     // Create a conversation profile
-    ConversationProfileManagement.createConversationProfileArticleSuggestion(
-        PROJECT_ID, CONVERSATION_PROFILE_DISPLAY_NAME, LOCATION, Optional.empty());
+    ConversationProfileManagement.createConversationProfile(
+        PROJECT_ID, CONVERSATION_PROFILE_DISPLAY_NAME, LOCATION);
     String output = bout.toString();
     assertThat(output).contains(NAME_PREFIX_IN_OUTPUT);
     conversationProfileName =
